+++ packages/apps/plugins/plugin-table/src/components/ObjectTable/hooks/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export * from './useObjects';
+export * from './useTables';
