+++ packages/apps/plugins/plugin-table/src/components/ObjectTable/hooks/useTables.ts
@@ -1,0 +1,14 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import { useMemo } from 'react';
+
+import { TableType } from '@braneframe/types';
+import { Filter } from '@dxos/echo-schema';
+import { type Space, useQuery } from '@dxos/react-client/echo';
+
+export const useTables = (space?: Space) => {
+  const tableFilter = useMemo(() => Filter.schema(TableType), []);
+  return useQuery<TableType>(space, tableFilter);
+};
